organize datacollector
