+++ minari/__init__.py
@@ -1,23 +1,20 @@
-from minari.minari_dataset import MinariDataset
-from minari.utils import (
-    combine_datasets,
-    create_dataset_from_buffers,
-    create_dataset_from_collector_env,
+from minari.data_collector.data_collector import (
+    DataCollectorV0,
+    EpisodeMetadataCallback,
+    StepDataCallback,
 )
+from minari.minari_dataset import MinariDataset
 from minari.storage.hosting import (
     download_dataset,
     list_remote_datasets,
     upload_dataset,
 )
-from minari.storage.local import (
-    delete_dataset,
-    list_local_datasets,
-    load_dataset
+from minari.storage.local import delete_dataset, list_local_datasets, load_dataset
+from minari.utils import (
+    combine_datasets,
+    create_dataset_from_buffers,
+    create_dataset_from_collector_env,
 )
-from minari.data_collector import (
-    DataCollectorV0,
-    EpisodeMetadataCallback,
-    StepDataCallback,
-)
+
 
 __version__ = "0.3.0"
