+++ minari/data_collector/__init__.py
@@ -1,0 +1,1 @@
+from minari.data_collector.data_collector import DataCollectorV0
